flight & mana regen
- changed mana to float
- added passive mana regen
- added flight using mana
*flight is toggled by a public bool canFly

-also tweaked some rigidbody and movements stats idk how they'd be tho

+++ Assets/Scripts/Manager Classes/UIManager.cs
@@ -31,12 +31,12 @@ public class UIManager : MonoBehaviour
         m_HealthSlider.value = health;
     }
 
-    public void SetMaxMana(int mana)
+    public void SetMaxMana(float mana)
     {
         m_ManaSlider.maxValue = mana;
     }
 
-    public void SetCurrentMana(int mana)
+    public void SetCurrentMana(float mana)
     {
         m_ManaSlider.value = mana;
     }
